Binominal fertig.

+++ src/aufgabenblatt4/InterfacePascal.java
@@ -4,7 +4,7 @@ package aufgabenblatt4;
 public abstract interface  InterfacePascal {
     
     
-    abstract int[] berechneZeile(int n);
+    abstract long[] berechneZeile(int n);
     
     abstract int getZaehler();
